Rewritten texture manager and texture cache. Vastly improved software renderer. Lots of changes, too many to list them here. Will post a detailed message on the mailing list.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2929 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csws/csgfxppl.cpp
@@ -192,9 +192,9 @@ void *PackDPFX (G3DPolygonDPFX &poly)
   PUT (bool, poly.use_fog);
   PUT (float, poly.inv_aspect);
   PUT (iTextureHandle *, poly.txt_handle);
-  PUT (float, poly.flat_color_r);
-  PUT (float, poly.flat_color_g);
-  PUT (float, poly.flat_color_b);
+  PUT (UByte, poly.flat_color_r);
+  PUT (UByte, poly.flat_color_g);
+  PUT (UByte, poly.flat_color_b);
 
 #undef PUT
 
@@ -217,9 +217,9 @@ void UnpackDPFX (void *Packed, G3DPolygonDPFX &poly)
   GET (bool, poly.use_fog);
   GET (float, poly.inv_aspect);
   GET (iTextureHandle *, poly.txt_handle);
-  GET (float, poly.flat_color_r);
-  GET (float, poly.flat_color_g);
-  GET (float, poly.flat_color_b);
+  GET (UByte, poly.flat_color_r);
+  GET (UByte, poly.flat_color_g);
+  GET (UByte, poly.flat_color_b);
 
 #undef GET
